# docs/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Movie API",
    "description": "A simple API for managing movie data, integrated with a MySQL database using Knex.js. This API allows listing, searching, adding, updating, and deleting movie records, and includes basic user authentication.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
    "/movies": {
      "get": {
        "tags": ["Movies"],
        "summary": "Get all movies",
        "description": "Retrieves a list of all movies with basic details. Limited to 100 results per the current implementation.",
        "responses": {
          "200": {
            "description": "A list of movies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": false },
                    "Message": { "type": "string", "example": "Success" },
                    "Movies": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieOverview"
                      }
                    }
                  }
                },
                "examples": {
                  "success": {
                    "summary": "Example Movie List",
                    "value": {
                      "Error": false,
                      "Message": "Success",
                      "Movies": [
                        {
                          "tconst": "tt0035423",
                          "primaryTitle": "Kate & Leopold",
                          "startYear": "2001",
                          "titleType": "movie"
                        },
                        {
                          "tconst": "tt0069049",
                          "primaryTitle": "The Other Side of the Wind",
                          "startYear": "2018",
                          "titleType": "movie"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "post": {
        "tags": ["Movies"],
        "summary": "Add a new movie",
        "description": "Adds a new movie record to the database. Only 'primaryTitle' and 'startYear' are required. Other fields will be auto-generated or defaulted by the server.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieCreateRequest"
              },
              "examples": {
                "minimalNewMovie": {
                  "summary": "Minimal Example: Title and Year",
                  "value": {
                    "primaryTitle": "The Great New Movie",
                    "startYear": "2025"
                  }
                },
                "fullNewMovie": {
                  "summary": "Full Example: All Fields Provided",
                  "value": {
                    "tconst": "tt9999999",
                    "titleType": "movie",
                    "primaryTitle": "A Brand New Film",
                    "originalTitle": "A Brand New Film",
                    "isAdult": 0,
                    "startYear": "2024",
                    "endYear": "",
                    "runtimeMinutes": "100",
                    "genres": "Drama,Thriller"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Movie added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "example": 32569367, "description": "The auto-generated ID for the new movie" },
                    "message": { "type": "string", "example": "Movie added successfully" },
                    "movie": { "$ref": "#/components/schemas/MovieFullDetails" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing required user-input fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Missing required field: primaryTitle" }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Movie with provided tconst already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Movie with this tconst (tt9999999) already exists. Please provide a unique tconst or omit it for automatic generation." }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/movies/search": {
      "get": {
        "tags": ["Movies"],
        "summary": "Search movies by title",
        "description": "Searches for movies where the primary title contains the given search term (case-insensitive partial match). Returns up to 50 results.",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "Part of the movie title to search for",
            "required": true,
            "schema": {
              "type": "string",
              "example": "wind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of matching movies",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": false },
                    "Message": { "type": "string", "example": "Success" },
                    "Movies": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieOverview"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing search term",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": true },
                    "Message": { "type": "string", "example": "Search term \"title\" is required." }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No movies found matching the search",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": false },
                    "Message": { "type": "string", "example": "No movies found matching your search." },
                    "Movies": { "type": "array", "items": {}, "example": [] }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/movies/data/{tconst}": {
      "get": {
        "tags": ["Movies"],
        "summary": "Get movie by IMDb ID from local database",
        "description": "Retrieves a single movie record by its tconst (IMDb ID) from the local MySQL database.",
        "parameters": [
          {
            "name": "tconst",
            "in": "path",
            "description": "IMDb ID of the movie to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "example": "tt0035423"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": false },
                    "Message": { "type": "string", "example": "Success" },
                    "Movies": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MovieFullDetails"
                      }
                    }
                  }
                },
                "examples": {
                  "singleMovie": {
                    "summary": "Example Single Movie",
                    "value": {
                      "Error": false,
                      "Message": "Success",
                      "Movies": [
                        {
                          "tconst": "tt0035423",
                          "titleType": "movie",
                          "primaryTitle": "Kate & Leopold",
                          "originalTitle": "Kate & Leopold",
                          "isAdult": 0,
                          "startYear": "2001",
                          "endYear": "",
                          "runtimeMinutes": "118",
                          "genres": "Comedy,Fantasy,Romance",
                          "id": 23041989
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Movie not found in local database",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": true },
                    "Message": { "type": "string", "example": "Movie with ID tt1234567 not found in local database" }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/movies/poster/{tconst}": {
      "get": {
        "tags": ["Movies"],
        "summary": "Get movie poster details by IMDb ID",
        "description": "Retrieves movie poster URL, title, and year from OMDb API for a given tconst. Requires the movie to exist in the local database first.",
        "parameters": [
          {
            "name": "tconst",
            "in": "path",
            "description": "IMDb ID of the movie to retrieve poster for",
            "required": true,
            "schema": {
              "type": "string",
              "example": "tt0035423"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Movie poster details fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoviePosterResponse"
                },
                "examples": {
                  "success": {
                    "summary": "Example Poster Response",
                    "value": {
                      "Error": false,
                      "Message": "Poster fetched successfully",
                      "Poster": "https://m.media-amazon.com/images/M/MV5BNjQ2MTE1NjY1OV5BMl5BanBnXkFtZTcwNzYxNzU4NA@@._V1_SX300.jpg",
                      "Title": "Kate & Leopold",
                      "Year": "2001"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Movie not found locally or poster not available from OMDb",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Error": { "type": "boolean", "example": true },
                    "Message": { "type": "string", "example": "Movie with ID tt1234567 not found in local database" }
                  }
                },
                "examples": {
                  "localNotFound": {
                    "summary": "Movie not in local DB",
                    "value": { "Error": true, "Message": "Movie with ID tt1234567 not found in local database" }
                  },
                  "omdbNotFound": {
                    "summary": "Poster not found on OMDb",
                    "value": { "Error": true, "Message": "Movie not found!" }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "MovieOverview": {
        "type": "object",
        "description": "Basic movie details returned by list and search endpoints. Required fields: tconst, primaryTitle, startYear, titleType.",
        "properties": {
          "tconst": { "type": "string", "description": "IMDb ID" },
          "primaryTitle": { "type": "string", "description": "Primary title of the movie" },
          "startYear": { "type": "string", "description": "Release year" },
          "titleType": { "type": "string", "description": "Type of title (e.g., movie, tvSeries)" }
        },
        "required": ["tconst", "primaryTitle", "startYear", "titleType"]
      },
      "MovieCreateRequest": {
        "type": "object",
        "description": "Minimal movie details required for creation (POST) endpoint. Other fields will be auto-generated or defaulted by the server.",
        "properties": {
          "primaryTitle": { "type": "string", "description": "The most common title of the movie", "example": "The New Era" },
          "startYear": { "type": "string", "description": "The release year of the movie", "example": "2025" },
          "tconst": { "type": "string", "description": "Optional IMDb ID (will be auto-generated if omitted)", "example": "tt1234567" },
          "titleType": { "type": "string", "description": "Type of title (e.g., movie, tvSeries)", "example": "movie" },
          "originalTitle": { "type": "string", "description": "The original title (defaults to primaryTitle if omitted)", "example": "The New Era" },
          "isAdult": { "type": "integer", "format": "int32", "description": "Indicates if the title is adult content (0 for false, 1 for true)", "example": 0 },
          "endYear": { "type": "string", "description": "End year for TV series (empty for movies)", "example": "" },
          "runtimeMinutes": { "type": "string", "description": "Primary runtime of the title, in minutes", "example": "90" },
          "genres": { "type": "string", "description": "Comma-separated list of genres (defaults to 'Unknown' if omitted)", "example": "Action,Adventure" }
        },
        "required": ["primaryTitle", "startYear"]
      },
      "MovieFullDetails": {
        "allOf": [
          { "$ref": "#/components/schemas/MovieCreateRequest" },
          {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "description": "Auto-incrementing internal database ID", "example": 23041989 }
            },
            "required": ["id"]
          }
        ],
        "description": "Complete movie details, including the internal ID, returned by the GET by tconst endpoint"
      },
...
      "MoviePosterResponse": {
        "type": "object",
        "description": "Response structure for fetching a movie poster.",
        "properties": {
          "Error": { "type": "boolean", "example": false },
          "Message": { "type": "string", "example": "Poster fetched successfully" },
          "Poster": { "type": "string", "format": "url", "description": "URL of the movie poster", "example": "https://m.media-amazon.com/images/M/MV5BNjQ2MTE1NjY1OV5BMl5BanBnXkFtZTcwNzYxNzU4NA@@._V1_SX300.jpg" },
          "Title": { "type": "string", "description": "Title of the movie", "example": "Kate & Leopold" },
          "Year": { "type": "string", "description": "Year of the movie", "example": "2001" }
        },
        "required": ["Error", "Message", "Poster", "Title", "Year"]
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "JWT token stored in an HttpOnly cookie"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "Error": { "type": "boolean", "example": true },
                "Message": { "type": "string", "example": "Error in MYSQL query or server failure" }
              }
            }
          }
        }
      }
    }
  }
}